requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiogram==3.2.0
aiohttp==3.9.1
psycopg[binary,pool]==3.1.13
pydantic-settings==2.1.0
openai==1.3.7
tiktoken==0.5.2
PyPDF2==3.0.1
python-docx==1.1.0
APScheduler==3.10.4
python-multipart==0.0.6
numpy==1.24.3
